requirements.txt
python-telegram-bot==21.8
fastapi==0.115.4
uvicorn[standard]==0.32.1
redis==5.2.0
httpx==0.27.2
pydantic==2.10.2
pydantic-settings==2.7.0
python-multipart==0.0.12
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
structlog==24.4.0
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-mock==3.12.0
python-dotenv==1.0.1
aiofiles==24.1.0
